TODO arrumar a selection_4

+++ src/Variable.java
@@ -1,14 +1,14 @@
 public class Variable {
 
     private char letter;
-    private int value;
+    private float value;
 
     public Variable(char letter, int value){
         this.letter = letter;
         this.value = value;
     }
 
-    public int getValue(){
+    public float getValue(){
         return this.value;
 
     }
